requirements.txt
aiogram
openai
python-dotenv
aiosqlite
